+++ lib/src/api/deserialization_context.dart
@@ -1,6 +1,6 @@
 import 'package:rdf_core/rdf_core.dart';
 import 'package:rdf_mapper/src/api/deserializer.dart';
-import 'package:rdf_mapper/src/api/node_reader.dart';
+import 'package:rdf_mapper/src/api/resource_reader.dart';
 
 /// Context for deserialization operations in the RDF mapping process.
 ///