+++ test/src/mappers/iri_mappers_test.dart
@@ -1,8 +1,8 @@
 import 'package:rdf_core/rdf_core.dart';
 import 'package:rdf_mapper/src/api/deserialization_context.dart';
 import 'package:rdf_mapper/src/api/deserializer.dart';
-import 'package:rdf_mapper/src/api/node_builder.dart';
-import 'package:rdf_mapper/src/api/node_reader.dart';
+import 'package:rdf_mapper/src/api/resource_builder.dart';
+import 'package:rdf_mapper/src/api/resource_reader.dart';
 import 'package:rdf_mapper/src/api/serialization_context.dart';
 import 'package:rdf_mapper/src/api/serializer.dart';
 import 'package:rdf_mapper/src/exceptions/deserialization_exception.dart';